+++ R/render.r
@@ -36,7 +36,7 @@ render_page <- function(pkg = ".", name, data, path = "", depth = NULL, quiet = 
 
   # The real location of 404.html is dynamic (#1129).
   # Relative root does not work, use the full URL if available.
-  if(identical(path, "404.html") && length(pkg$meta$url)){
+  if (path == "404.html" && length(pkg$meta$url)) {
     data$site$root <- paste0(pkg$meta$url, "/")
   }
 
@@ -56,6 +56,10 @@ render_page <- function(pkg = ".", name, data, path = "", depth = NULL, quiet = 
   components <- purrr::map(templates, render_template, data = data)
   components <- purrr::set_names(components, pieces)
   components$template <- name
+
+  if (path == "404.html" && !is.null(pkg$meta$url)) {
+    components$navbar <- tweak_navbar_links(components$navbar, pkg = pkg)
+  }
 
   # render complete layout
   template <- find_template(
@@ -79,7 +83,7 @@ data_template <- function(pkg = ".", depth = 0L) {
     paste(collapse = ", ")
 
   # Force inclusion so you can reliably refer to objects inside yaml
-  # in the moustache templates
+  # in the mustache templates
   yaml <- purrr::pluck(pkg, "meta", "template", "params", .default = list())
   yaml$.present <- TRUE